Category->DEPARTMENT_ID

+++ src/main/java/com/divudi/bean/pharmacy/GoodsReturnController.java
@@ -257,7 +257,7 @@ public class GoodsReturnController implements Serializable {
             //System.err.println("Billed " + rBilled);
             //System.err.println("Cancelled " + rCacnelled);
             //System.err.println("Net " + netQty);
-            retPh.setQtyInUnit((float) (grnPh.getQtyInUnit() - netQty));
+            retPh.setQtyInUnit((double) (grnPh.getQtyInUnit() - netQty));
 
             List<Item> suggessions = new ArrayList<>();
             Item item = bi.getItem();
